Implemented barebones ProductDetail componenets
