--- client/src/components/Overview/AddToCart.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import { BsStar } from 'react-icons/bs';
-
-const AddToCart = () => (
-  <div className='add-container'>
-    <button>SELECT SIZE</button>
-    <button>QUANTITY: 42</button>
-    <button>ADD TO CART</button>
-    <button><BsStar /></button>
-  </div>
-)
-
-export default AddToCart;
