--- client/src/components/Overview/GalleryEntry.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-const GalleryEntry = () => (
-  <div className='g-entry'>
-  </div>
-)
-
-export default GalleryEntry;
