--- client/src/components/Overview/ImageGallery.jsx
@@ -1,37 +1,0 @@
-import React from 'react';
-import { IoMdArrowDropdown } from 'react-icons/io';
-import './overview.css';
-import GalleryEntry from './GalleryEntry.jsx'
-import ProductView from './ProductView.jsx'
-
-class imageGallery extends React.Component {
-  constructor(props) {
-    super(props)
-
-    this.state = {
-
-    }
-  }
-
-
-  render() {
-    return (
-      <div className='image-container'>
-        <div className='g-container'>
-          <GalleryEntry />
-          <GalleryEntry />
-          <GalleryEntry />
-          <GalleryEntry />
-          <GalleryEntry />
-          <IoMdArrowDropdown className='arrow' />
-        </div>
-        <div>
-          <ProductView />
-        </div>
-      </div>
-    )
-  }
-}
-
-
-export default imageGallery;
