--- client/src/components/Overview/ProductOverview.jsx
@@ -1,20 +1,0 @@
-import React from 'react';
-import { BsCheckLg } from 'react-icons/bs';
-
-const ProductOverview = () => (
-  <div>
-    <div>
-      <h2>Lorem Ipsum the <br /> printing and typesetting industry.</h2>
-      <p></p>
-      <p></p>
-    </div>
-    <div>
-      <p><BsCheckLg />GMO</p>
-      <p><BsCheckLg />GMO</p>
-      <p><BsCheckLg />GMO</p>
-      <p><BsCheckLg />GMO</p>
-    </div>
-  </div>
-)
-
-export default ProductOverview;
