--- client/src/components/Overview/Star.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import { BsStarFill, BsStarHalf, BsStar } from 'react-icons/bs';
-
-const ProductView = () => (
-  <div>
-    <BsStarFill className='star' />
-    <BsStarFill className='star' />
-    <BsStarFill className='star' />
-    <BsStarHalf className='star' />
-    <BsStar className='star' />
-    <a href=''>Read all reviews</a>
-  </div>
-)
-
-export default ProductView;
