--- client/src/components/Overview/Style.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Style = () => (
-  <div className='style-entry'></div>
-)
-
-export default Style;
