+++ client/src/components/ProductDetail/AddToCart.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { TiStarFullOutline } from 'react-icons/ti';
+
+const AddToCart = () => (
+  <div className='add-container'>
+    <button>SELECT SIZE</button>
+    <button>QUANTITY: 42</button>
+    <button>ADD TO CART</button>
+    <button><TiStarFullOutline /></button>
+  </div>
+)
+
+export default AddToCart;
