+++ client/src/components/ProductDetail/GalleryEntry.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+const GalleryEntry = () => (
+  <div className='g-entry'>
+  </div>
+)
+
+export default GalleryEntry;
