+++ client/src/components/ProductDetail/Star.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { TiStarFullOutline, TiStarHalfOutline, TiStarOutline } from 'react-icons/ti';
+
+const ProductView = () => (
+  <div>
+    <TiStarFullOutline className='star' />
+    <TiStarFullOutline className='star' />
+    <TiStarFullOutline className='star' />
+    <TiStarHalfOutline className='star' />
+    <TiStarOutline className='star' />
+    <a href=''>Read all reviews</a>
+  </div>
+)
+
+export default ProductView;
