+++ client/src/components/ProductDetail/Style.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Style = () => (
+  <div className='style-entry'></div>
+)
+
+export default Style;
