+++ client/src/components/ProductDetail/overview.css
@@ -1,0 +1,119 @@
+.main-container {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+/* ----------------- Image Gallery ----------------- */
+
+.image-container {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: row;
+  background-color: #94B49F;
+}
+
+.g-container {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+}
+
+.g-entry {
+  background-color: #fff;
+  width: 50px;
+  margin: 20px;
+  height: 50px;
+  transition: all .5s ease;
+}
+
+.g-entry:hover {
+  cursor: pointer;
+  border-bottom: 5px solid #666;
+}
+
+.pv-container {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin: 10px;
+}
+
+.arrow {
+  color: #fff;
+  cursor: pointer;
+  font-size: 20px;
+}
+
+.expand {
+  color: #fff;
+  cursor: pointer;
+  font-size: 30px;
+  align-self: flex-start;
+}
+
+/* ----------------- Product Information ----------------- */
+
+.info-container {
+  margin: 0 30px;
+  width: 250px;
+}
+
+.star {
+  margin: 0 2px;
+}
+
+.style-title {
+  display: inline-flex;
+}
+
+.style-entry {
+  border-radius: 30px;
+  border: solid 0.5px #333;
+  height: 50px;
+  width: 50px;
+  margin: 5px;
+  background-color: #94B49F;
+}
+
+.style-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-wrap: wrap;
+}
+
+/* ----------------- Add To Cart ----------------- */
+
+.add-container {
+  margin: 30px 0;
+}
+
+.add-container > button {
+  padding: 10px 10px;
+  margin: 5px 7px;
+}
+
+
+/* ----------------- Product Overview ----------------- */
+.prodview-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: row;
+  margin: 40px 400px;
+}
+
+.prodview-text {
+  width: 100%;
+}
+
+.prodview-line {
+  margin: 0 40px;
+  background-color: #777;
+  border-radius: 7px;
+  height: 150px;
+  width: 4px;
+}
